pyproject.toml
[project]
name = "circom-analyzer"
version = "0.0.1"
authors = [
  { name="Youwei Zhong", email="zhongyouwei@sjtu.edu.cn" },
  { name="Yanju Chen", email="yanju@veridise.com" },
]
description = "Circom to C translator"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "antlr4-python3-runtime==4.13.1",
    "antlr4-tools",
    "tabulate>=0.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
circom-analyzer = "translator.circom.run:run"